fix(fleet): one quiet runner should not hold up startup
DiscoverInventory probed resources one at a time with no per-resource
deadline, so a server that completed its TCP handshake and then went
quiet had nothing bounding it. At boot that stalls the process; on a
refresh it stalls every resource queued behind it, and the fleet learns
nothing about endpoints that were answering perfectly well. It is the
same failure the stream-idle guard closes during generation, reached one
loop earlier and with more to lose, because at boot there is no fleet
yet to route around the gap.

Probes now run concurrently, each under its own deadline. Results are
written by index rather than appended, so the inventory keeps catalog
order however the probes interleave — order is part of the contract for
anything downstream that reads the list, and making it depend on which
runner answered first would trade one nondeterminism for another.

Extracting probeResource is what made that possible. The old loop body
appended to the shared inventory from inside its provider branches, so
the work could not move off the goroutine that owned the slice; the
function returns its result and a flag for whether the resource belongs
in the inventory at all, since a provider with nothing to discover
(anthropic) is not a failure, it is silent.

The size guardrail objected afterward, and it had a point: the file was
doing two jobs. Discovery stays in inventory.go and the merge half —
deciding what a probe result may change about the configured catalog —
moves to inventory_merge.go. Metrics are back at baseline without a bump.

One note for the next person writing a test here: httptest.Close waits
for in-flight handlers, so a deliberately stalled handler has to be
released before the server is closed. Defers run LIFO, which means the
release has to be deferred last, and getting that backwards deadlocks
the test rather than failing it.

Refs #1406.
